Revert "[wlan] Use new boringssl Rust bindings"
This reverts commit 5297e823353a577881da90d62e2fb976aba6142b.

Reason for revert: b/395499454

Original change's description:
> [wlan] Use new boringssl Rust bindings
>
> There are new, canonical Rust bindings now shipped in boringssl.git.
> They are still backed by a bindgen-generated file produced in-tree,
> which now uses the vendored prebuilt `bindgen`, accounting for the
> i32-vs-u32 difference seen.
>
> For parity with the old bindings, we replace the dependency with a
> flavour that only depends on libcrypto.
>
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1200410
> Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
> Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>

Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1202804

+++ src/connectivity/wlan/lib/sae/src/boringssl.rs
@@ -332,7 +332,7 @@ pub enum EcGroupId {
 }
 
 impl EcGroupId {
-    fn nid(&self) -> i32 {
+    fn nid(&self) -> u32 {
         match self {
             EcGroupId::P256 => NID_X9_62_prime256v1,
             EcGroupId::P384 => NID_secp384r1,
@@ -362,7 +362,7 @@ impl Drop for EcGroup {
 
 impl EcGroup {
     pub fn new(id: EcGroupId) -> Result<Self, Error> {
-        ptr_or_error(unsafe { EC_GROUP_new_by_curve_name(id.nid()) }).map(Self)
+        ptr_or_error(unsafe { EC_GROUP_new_by_curve_name(id.nid() as i32) }).map(Self)
     }
 
     /// Returns the prime and curve parameters that constitute this group.
